update the POM
added Junit for the new functionality that we want to compartmentalize
on a different section.

separate the Clipboard Listening functionality into its own section in
order to extend and allow better handling. also to asurre responsiveness
when big chunks of data are read from the clipboard we setup to run as
its own "service thread"
 Changes to be committed:
	modified:   pom.xml
	new file:   src/main/java/com/aeongames/edi/utils/Clipboard/ClipBoardListener.java
	new file:   src/main/java/com/aeongames/edi/utils/Clipboard/FlavorHandler.java
	new file:   src/main/java/com/aeongames/edi/utils/Clipboard/FlavorProcessor.java
	new file:   src/main/java/com/aeongames/edi/utils/ThreadUtils/StopSignalProvider.java
	modified:   src/main/java/com/aeongames/edi/utils/error/ErrorData.java
	modified:   src/main/java/com/aeongames/edi/utils/error/LoggingHelper.java
	new file:   src/main/java/com/aeongames/sync/SignalWaiter.java
	new file:   src/test/java/com/aeongames/edi/tests/TestClipboardListenings.java