Change the approach to the CLipboard listener as previous method spanned several threads with no real benefit in performance nor responsiveness.
new approach creates a new thread as a Daemon that run on the background 
waiting on a queue of work to be done 

when a new change is detected on the Clipboard the threads responds (if not alredy
processing work) and handle the change on the clipboard.

furthermore the new approach allow a modular approach to the data handling
in a way that several classes can be register to handle the data, note however
that this approach is intended not for notification but for "first come first serve"
data handling and thus when a class that can handle the data from the 
clipboard reports success no other will be called.
some work can be made to adapt to a different approach but that is not 
the goal of this tool.